+++ tests/test_wordEncoding.txt
